+++ lib/attic/files_and_folders.dart
@@ -1,6 +1,16 @@
+/// FOLDERS
+
 const String jsonIntermediaryDirectory = "/json";
+
 const String votesDirectory = "/votes";
 const String amendementsDirectory = "/amendements";
 const String docsLegisDirectory = "/docs_legis";
 const String dossierParlementaireDirectory = "/dossierParlementaire";
 const String documentDirectory = "/document";
+
+/// FILES
+
+const String votesZipFile = "/votes.zip";
+const String amendementsZipFile = "/amendements.zip";
+const String dossierParlementaireZipFile = "/dossiers.zip";
+const String deputesCsvFile = "/deputes.csv";